+++ internal/kubernetes/kubeconfig/kubeconfig_resource.go
@@ -15,7 +15,7 @@ import (
 	"k8s.io/client-go/tools/clientcmd"
 	k8sApi "k8s.io/client-go/tools/clientcmd/api"
 
-	swagger "github.com/crusoecloud/client-go/swagger/v1alpha5"
+	swagger "github.com/crusoecloud/client-go/swagger/v1"
 	"github.com/crusoecloud/terraform-provider-crusoe/internal/common"
 )